commit to version 1.2

+++ provinder_8081/src/main/java/com/grace/it/service/DeptManagerService.java
@@ -21,7 +21,7 @@ public interface DeptManagerService {
 
     public Integer updateByCondition(DeptManager deptManager);
 
-    public Long insert(DeptManager deptManager);
+    public Integer insert(DeptManager deptManager);
 
     public Integer deleteByEmpNo(Long empNo);
 }
